requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
python-dotenv==1.0.1
bcrypt==4.1.2
docker==7.0.0
boto3==1.34.23
mysql-connector-python==8.3.0

#Run the command 
pip freeze > requirements.txt